pyproject.toml
[project]
name = "check-ds"
version = "0.1.0"
description = "Веб-сервис проверки дополнительных соглашений к госконтрактам"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-multipart>=0.0.6",
    "python-docx>=1.1.0",
    "openpyxl>=3.1.0",
    "xlrd>=2.0.0",
    "pandas>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "aiosqlite>=0.19.0",
    "rarfile>=4.1",
    "jinja2>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.26.0",
]
win = [
    "pywin32>=306",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["api", "core", "db", "schemas", "services"]
